requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
Werkzeug==1.0.1
